+++ main/gameseq.c
@@ -815,7 +815,7 @@ void set_sound_sources()
 						vms_vector pnt;
 
 						compute_center_point_on_side(&pnt,seg,sidenum);
-						digi_link_sound_to_pos(sn,segnum,sidenum,&pnt,1, F1_0/2);
+						digi_link_sound_to_pos(sn,(short) segnum,(short) sidenum,&pnt,1, F1_0/2);
 
 					}
 		}
@@ -851,7 +851,7 @@ void create_player_appearance_effect(object *player_obj)
 		effect_obj->orient = player_obj->orient;
 
 		if ( Vclip[VCLIP_PLAYER_APPEARANCE].sound_num > -1 )
-			digi_link_sound_to_object( Vclip[VCLIP_PLAYER_APPEARANCE].sound_num, effect_obj-Objects, 0, F1_0);
+			digi_link_sound_to_object( Vclip[VCLIP_PLAYER_APPEARANCE].sound_num, (short) (effect_obj-Objects), 0, F1_0);
 	}
 }
