+++ main/multibot.c
@@ -887,11 +887,11 @@ multi_do_create_robot(char *buf)
 	// Play effect and sound
 
 	compute_segment_center(&cur_object_loc, &Segments[robotcen->segnum]);
-	obj = object_create_explosion(robotcen->segnum, &cur_object_loc, i2f(10), VCLIP_MORPHING_ROBOT);
+	obj = object_create_explosion((short) robotcen->segnum, &cur_object_loc, i2f(10), VCLIP_MORPHING_ROBOT);
 	if (obj)
 		extract_orient_from_segment(&obj->orient, &Segments[robotcen->segnum]);
 	if (Vclip[VCLIP_MORPHING_ROBOT].sound_num > -1)
-		digi_link_sound_to_pos( Vclip[VCLIP_MORPHING_ROBOT].sound_num, robotcen->segnum, 0, &cur_object_loc, 0, F1_0 );
+		digi_link_sound_to_pos( Vclip[VCLIP_MORPHING_ROBOT].sound_num, (short) robotcen->segnum, 0, &cur_object_loc, 0, F1_0 );
 
 	// Set robot center flags, in case we become the master for the next one
 
@@ -976,9 +976,9 @@ multi_do_boss_actions(char *buf)
 				vm_vec_sub(&boss_dir, &Objects[Players[pnum].objnum].pos, &boss_obj->pos);
 				vm_vector_2_matrix(&boss_obj->orient, &boss_dir, NULL, NULL);
 
-				digi_link_sound_to_pos( Vclip[VCLIP_MORPHING_ROBOT].sound_num, teleport_segnum, 0, &boss_obj->pos, 0 , F1_0);
+				digi_link_sound_to_pos( Vclip[VCLIP_MORPHING_ROBOT].sound_num, (short) teleport_segnum, 0, &boss_obj->pos, 0 , F1_0);
 				digi_kill_sound_linked_to_object( boss_obj-Objects);
-				digi_link_sound_to_object2( SOUND_BOSS_SHARE_SEE, boss_obj-Objects, 1, F1_0, F1_0*512 );	//	F1_0*512 means play twice as loud
+				digi_link_sound_to_object2( SOUND_BOSS_SHARE_SEE, (short) (boss_obj-Objects), 1, F1_0, F1_0*512 );	//	F1_0*512 means play twice as loud
 				Ai_local_info[boss_obj-Objects].next_fire = 0;
 
 				if (boss_obj->ctype.ai_info.REMOTE_OWNER == Player_num)
