+++ main/physics.c
@@ -944,9 +944,9 @@ save_p1 = *fq.p1;
 				wall_part = vm_vec_dot(&moved_v,&hit_info.hit_wallnorm);
 
 				if (wall_part != 0 && moved_time>0 && (hit_speed=-fixdiv(wall_part,moved_time))>0)
-					collide_object_with_wall( obj, hit_speed, WallHitSeg, WallHitSide, &hit_info.hit_pnt );
+					collide_object_with_wall( obj, hit_speed, (short) WallHitSeg, (short) WallHitSide, &hit_info.hit_pnt );
 				else
-					scrape_object_on_wall(obj, WallHitSeg, WallHitSide, &hit_info.hit_pnt );
+					scrape_object_on_wall(obj, (short) WallHitSeg, (short) WallHitSide, &hit_info.hit_pnt );
 
 				Assert( WallHitSeg > -1 );				
 				Assert( WallHitSide > -1 );
